pil: modem: assign/free memory to/from modem before/after MBA bootup
When a GVM with a modem is shutdown, MSS region is assigned back to APPS.
So during the GVM restart procedure, we need to assign it to modem before
starting MBA bootup since in secure-boot device, MBA will scribble the MSS
region. After MBA bootup, we assign back such region to APPS again.

Change-Id: Icaea84013bff59467a795a84a301b7b4aa3c22f9
Signed-off-by: Yong Ding <yongding@codeaurora.org>