msm: ekms: Remove the wait for completion of eDRM probe
Main DRM & eDRM driver has asynchronous probes. In order to ensure that
both the driver's probe has completed there is a wait for completion for
respective driver is executed. There is an issue where the
wait for the eDRM driver blocks the eDRM driver deferred probe call.

This change remove wait in the eDRM driver for completion of its probe
as eDRM driver is always initialized after main DRM & necessary clock
voting would also be done in main DRM.

Change-Id: I422419d381ad3d0361fb80f3b2b9d176203a9342
Signed-off-by: Rahul Sharma <rahsha@codeaurora.org>