adv7481: Add interrupt work arounds
The generic INT_RAW_STATUS register is sometimes not exactly
synchronized with the individual interrupt registers. Hence
avoid relying on this register in interrupt processing and check
the individual interrupt registers. Also schedule the same worker
thread again to recheck interrupt registers and handle them if
the INT_RAW_STATUS is set at the end of any interrupt processing.

Suggested-by: Resmi Rajendran
Change-Id: I29428301d7f132331adeaede0c51bdf5eb2f86f9
Signed-off-by: Suprith Malligere Shankaregowda <supgow@codeaurora.org>