drm/msm/dsi: add builtin bridge pos support in dsi panel
Builtin DSI bridge is the first bridge in the bridge chain by
default. Move the builtin bridge to other position if
qcom,mdss-dsi-builtin-bridge-pos is set in dsi panel devicetree.
This is useful when external bridge has dependency on DSI clock
at pre-enable stage and also for changing the dsi initialization
sequence.

Change-Id: I1bfc84ef4924ff18ce29ee467645b859d5c83d9a
Acked-by: Santosh Ramani <c_santra@qti.qualcomm.com>
Signed-off-by: Rahul Sharma <rahsha@codeaurora.org>