usb: host: xhci-ring: Add NULL check for xhci->devs
In fast connect and disconnect case of a peripheral, it is
possible that a pending transfer tries to access xhci->devs of
that peripheral slot as a part of xhci_triad_to_transfer_ring,
but the xhci->devs[slot_id] has already been freed and set to
NULL in xhci_free_virt_device as a part of disconnect handling.
Fix this NULL pointer dereference by adding a check.

Change-Id: Id1e400505e58445084b33ef56a62fa4d6c7e4e15
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>