drm/msm/dsi: fix bridge chip initialization order
This change is done for fixing the reboot issue observed
when split dsi is enabled.

In split DSI configuration,Bridge chips were initialized
in incorrect order because the builtin_bridge_pos flag was
not handled correctly for the respective bridge chips
associated with the DSI display.

Fix is to handle the builtin_bridge_pos flag correctly for bridge
chips associated with DSI displays for all the display
topology of DSI interface.

Change-Id: I4acd3024056e6ff5f0e9dc9bb3efd3c5ff4cb900
Acked-by: Santosh Ramani <c_santra@qti.qualcomm.com>
Signed-off-by: Rahul Sharma <rahsha@codeaurora.org>